+++ __helpers/handleThrowError.ts
@@ -1,3 +1,0 @@
-export const handleThrowError = (message: string): void => {
-  throw new Error(message);
-};