+++ CHANGELOG.md
@@ -23,13 +23,55 @@ All notable changes to this project will be documented in this file.
 - anything you're working on
 --->
 
+## [1.1.35] - 2020-03-05
+
+Changed the CTALink and Button interfaces to not share unnecessary props
+
+### Removed
+
+- removed handleClick, isSubmit, isLoading, url from the IButton interface
+
+### Added
+
+- added handleClick, isSubmit, isLoading to the IButtonProps interface
+- added url to the ICTALink interface
+
+## [1.1.34] - 2020-03-04
+
+### Fixed
+
+- fixed button styles for grey and ghost props
+
+### Changed
+
+- form now validates on blur
+
+## [1.1.33] - 2020-03-03
+
+### Changed
+
+- form now validates on blur
+
+## [1.1.32] - 2020-03-03
+
+### Changed
+
+- changed build step script to "publish"
+
+## [1.1.31] - 2020-03-03
+
+### Added
+
+- added build step to publish
+
 ## [1.1.30] - 2020-03-03
 
 ### Removed
 
 - removed hasValidation prop on all inputs in favour of using validationRules as an indicator
+- removed old **comoponents and **helpers folders
 
-### Added
+### Changed
 
 - now passing reset prop from react hook form to all children of Form